backend/requirements.txt
# Backend - Security Verification System (Python 3.8+)
python-dotenv==1.0.0
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.5.1
PyMySQL==1.1.0
PyJWT==2.9.0
bcrypt==4.1.1
face-recognition==1.3.0
flask-swagger-ui==4.11.1
Pillow==10.1.0
requests==2.31.0
Werkzeug==3.0.1
